refactor: single source of truth for domain<->proto conversion (internal/protoconv)
The same converters were maintained in triplicate — internal/server
(convert.go / mutations.go / config.go), the TUI (client.go) and the CLI
(automation.go / jobs.go) — because the depguard client boundary forbids
clients from importing the server. Every new Trigger/FleetSettings field
had to be added in up to 6 places, and the copies had already drifted:
the CLI's trigger converters lost the Disabled field, so any
`fleet trigger`/`fleet agent` write silently re-enabled every disabled
trigger in the fleet (the drift predicted when Trigger.Disabled shipped).

internal/protoconv now holds one converter per type per direction. Like
configutil it sits outside the depguard client rule: pure mapping over
DTO shapes, no state I/O. The package's reflection round-trip tests fill
every field of every domain struct and require deep equality, so a field
added to a domain struct without teaching its converter fails the suite
automatically — plus exhaustive enum, tri-state presence, and nil/empty
convention tests consolidated from the three per-surface test files.

Behavioral fix: CLI automation writes no longer drop Trigger.Disabled.
Deliberate divergence kept: ConfigFromProto takes a base — the server's
SetConfig write path passes a zero Config (never mask a cleared field),
the TUI read path passes DefaultConfig (absent fields render defaults).
